feat(lvnet): strike the old netlist format -- lvnet is the only text surface
The #70 lvnet work added the graph-native `build_netlist_from_graph` +
`render_lvnet` but LEFT the old Operation-based `render_netlist` ASCII
renderer in place as a separate `--format netlist` and the git-textconv
default. That was never the intent: netlist -> lvnet. lvnet is strictly
better (stable BD-uid identity, minimal-load-friendly, type-rehydratable);
no one relies on the old dataflow-netlist body.

Struck:
- `describe --format` drops `netlist` (choices are now text / lvnet / json);
  `--format lvnet` is the netlist text surface.
- `setup --git-textconv` now defaults to (and only offers) `lvnet`.
- Deleted `render_netlist` + `component_line` + their exclusive gamma/mu/eta
  render call graph, and the now-dead `_repo_relative_path` CLI helper.
- Renamed `netlist_render_deprecated.py` -> `netlist_diff_helpers.py`: it now
  holds ONLY the shared netlist-diff text helpers (`ambiguous_bares`,
  `instance_line`, `scope_header`) that `diff.py` still imports.
- Removed `tests/test_netlist.py` (the old-renderer suite) and the
  `--format netlist` cases in `tests/test_cli_describe_lvnet.py`.

NOT touched: the `build_netlist` (Operation) BUILDER stays -- diff, the json
surface, and node/vi-properties still consume it. Migrating those onto
`build_netlist_from_graph` (and rebuilding their Operation-based `_StubGraph`
fixtures) is the remaining Operation-removal work; diff keeps working here
because its dependencies were all preserved.
