+++ tests/test_cli_describe_lvnet.py
@@ -1,16 +1,13 @@
 """Tests for `lvkit describe --format lvnet` (Element 5: wiring `render_lvnet`
-into the `describe` CLI) and the `--format netlist` deprecation notice.
+into the `describe` CLI).
 
 Uses the same real corpus VI as `tests/test_render_lvnet.py`'s golden fixture
 (`loadTestsFromTestCase.vi`, JKI VI-Tester sample) so the CLI's output can be
 checked against the underlying `render_lvnet`/`build_netlist_from_graph`
 functions directly, in-process -- not just "the subprocess didn't crash".
 
-`--format netlist` is exercised too, ONLY to confirm this change left its
-bytes untouched (byte-for-byte against `tests/test_netlist.py`'s own
-machinery would duplicate that suite; here we just diff the subprocess's
-stdout against `render_netlist` called directly) and that the new stderr
-deprecation notice appears without touching stdout.
+(The OLD `--format netlist` was struck -- lvnet is the only netlist text
+surface now; `--format` no longer accepts `netlist`.)
 """
 
 from __future__ import annotations
@@ -24,10 +21,8 @@ import pytest
 
 from lvkit.graph.core import InMemoryVIGraph
 from lvkit.graph.netlist import (
-    build_netlist,
     build_netlist_from_graph,
     render_lvnet,
-    render_netlist,
 )
 from lvkit.load_mode import LoadMode
 
@@ -102,30 +97,6 @@ def test_format_lvnet_verbose_matches_render_lvnet_and_has_uses_and_types() -> N
     assert result.stdout == expected + "\n"
     assert "  uses :" in result.stdout
     assert "types :" in result.stdout
-
-
-@pytest.mark.needs_samples
-def test_format_netlist_deprecation_notice_on_stderr() -> None:
-    _require_golden()
-    result = _run_describe("--format", "netlist")
-    assert result.returncode == 0, result.stderr
-    assert "deprecated" in result.stderr
-    assert "lvnet" in result.stderr
-
-
-@pytest.mark.needs_samples
-def test_format_netlist_output_bytes_unchanged() -> None:
-    """--format netlist must still render the OLD render_netlist body,
-    byte-for-byte -- this change must not alter it (git textconv depends on
-    it staying stable)."""
-    _require_golden()
-    graph, vi_name = _load()
-    module = build_netlist(graph, vi_name)
-    expected = render_netlist(module, display_name=graph.vi_display_name(vi_name))
-
-    result = _run_describe("--format", "netlist")
-    assert result.returncode == 0, result.stderr
-    assert result.stdout == expected + "\n"
 
 
 @pytest.mark.needs_samples
